Move consolidations tests
