+++ beacon-chain/core/requests/BUILD.bazel
@@ -33,10 +33,14 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["withdrawals_test.go"],
+    srcs = [
+        "consolidations_test.go",
+        "withdrawals_test.go",
+    ],
     deps = [
         ":go_default_library",
         "//beacon-chain/state:go_default_library",
+        "//beacon-chain/state/state-native:go_default_library",
         "//config/params:go_default_library",
         "//consensus-types/primitives:go_default_library",
         "//encoding/bytesutil:go_default_library",
